requirements.txt
python-decouple
django